Squashfs: fix corruption checks in squashfs_lookup()
The dir_count and size fields when read from disk are sanity
checked for correctness.  However, the sanity checks only check the
values are not greater than expected.  As dir_count and size were
incorrectly defined as signed ints, this can lead to corrupted values
appearing as negative which are not trapped.

Change-Id: I95cf6a9401b64b8f6dae5386a62efb34b34f6ab0
Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>