lis3dh: Only transition FIFO mode through bypass if FIFO was full
FIFO data collection stops if FIFO is full, so the application note
suggests to transition FIFO mode through bypass to enable data
collection again. This is only necessary if the FIFO actually was full,
though, not if just the watermark level was reached. The unneeded mode
transition introduced bogus entries into the FIFO.

Change-Id: I4be65dfdcee88e231da29cdcd1d3199be83da7f9